requirements.txt
streamlit
firebase-admin
google-cloud-firestore
google-cloud-storage
google-genai
fsspec
gcsfs
nest_asyncio
pandas
matplotlib
seaborn